What did you RSpec?
A comparison between the behavior of a small fix-based script (which became the r_spec gem) and rspec .

As a result, with r_spec the build is passing, while with rspec it is failing. Despite two separate contexts, rspec was not able to evaluate some code in isolation to prevent side effects.